+++ src/views/LostItems/LostItemsListContainer/LostItemsListContainer.js
@@ -114,7 +114,7 @@ class LostItemsListContainer extends React.Component {
       .catch((e) => {
         let message;
 
-        if (e.status === STATUS_CODES.unprocessableEntity) {
+        if (e.status === STATUS_CODES.UNPROCESSABLE_ENTITY) {
           message = <FormattedMessage
             id="ui-users.lostItems.notification.billedBefore"
             values={{ patronName }}
@@ -157,7 +157,7 @@ class LostItemsListContainer extends React.Component {
       .catch((e) => {
         let message;
 
-        if (e.status === STATUS_CODES.unprocessableEntity) {
+        if (e.status === STATUS_CODES.UNPROCESSABLE_ENTITY) {
           message = <FormattedMessage
             id="ui-users.lostItems.notification.billedBefore"
             values={{ patronName }}
@@ -289,7 +289,6 @@ class LostItemsListContainer extends React.Component {
                     </Button>
                     <Filters
                       activeFilters={activeFilters.state}
-                      resources={resources}
                       onChangeHandlers={getFilterHandlers()}
                       resultOffset={resultOffset}
                     />