+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * Add STATUS FILTER to LOST ITEMS REQUIRING ACTUAL COST processing page. Refs UIU-2748.
 * Correctly handle removing all permissions from given user. Fixes UIU-2822.
+* Add "Status" filter to LOST ITEMS REQUIRING ACTUAL COST processing page. Refs UIU-2748.
+* Fix problem with remaining amount (not shown correct value after filling the payment amount). Refs UIU-2812.
+* Add "Status" column to LOST ITEMS REQUIRING ACTUAL COST processing page. Refs UIU-2772.
+* Add "Lost items requiring actual cost" to Actions dropdown in User details record. Refs UIU-2810.
 
 ## [9.0.0](https://github.com/folio-org/ui-users/tree/v9.0.0) (2023-02-20)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v8.1.0...v9.0.0)